requirements.txt
anthropic>=0.40.0
feedparser>=6.0.11
jinja2>=3.1.4
pyyaml>=6.0.2
requests>=2.32.0
python-dateutil>=2.9.0
